vaishnavi project/requirements.txt
streamlit
pymupdf
sentence-transformers
faiss-cpu
huggingface_hub
python-dotenv
transformers
